+++ share/classes/grade.class.php
@@ -69,13 +69,16 @@ class Grade {
      * @return mixed 
      */
     public function add(){
-        $db = DB::prepare('SELECT COUNT(id) FROM grade WHERE grade = ?');
-        $db->execute(array($this->grade));
-        if($db->fetchColumn() >= 1) { 
-            return 'Diesen Klassennamen gibt es bereits.';
-        } else {
-            $db = DB::prepare('INSERT INTO grade (grade,description,creator_id,institution_id) VALUES (?,?,?,?)');
-            return $db->execute(array($this->grade, $this->description, $this->creator_id, $this->institution_id));
+        global $USER;
+        if (checkCapabilities('grade:add', $USER->role_id)){
+            $db = DB::prepare('SELECT COUNT(id) FROM grade WHERE grade = ?');
+            $db->execute(array($this->grade));
+            if($db->fetchColumn() >= 1) { 
+                return 'Diesen Klassennamen gibt es bereits.';
+            } else {
+                $db = DB::prepare('INSERT INTO grade (grade,description,creator_id,institution_id) VALUES (?,?,?,?)');
+                return $db->execute(array($this->grade, $this->description, $this->creator_id, $this->institution_id));
+            }
         }
     }
     
@@ -84,8 +87,11 @@ class Grade {
      * @return boolean 
      */
     public function update(){
-        $db = DB::prepare('UPDATE grade SET grade = ?, description = ?, creator_id = ? WHERE id = ?');
-        return $db->execute(array($this->grade, $this->description, $this->creator_id, $this->id));
+        global $USER;
+        if (checkCapabilities('grade:update', $USER->role_id)){
+            $db = DB::prepare('UPDATE grade SET grade = ?, description = ?, creator_id = ? WHERE id = ?');
+            return $db->execute(array($this->grade, $this->description, $this->creator_id, $this->id));
+        }
     }
     
     /**
@@ -93,23 +99,25 @@ class Grade {
      * @return mixed 
      */
     public function delete($creator_id = null){
-        if ($creator_id != null) { // if function is called by request-php --> required by checkCapabilities()
+        /*if ($creator_id != null) { // if function is called by request-php --> required by checkCapabilities()
             $user = new USER();
 
             $user->load('id', $creator_id);
             $role_id = $user->role_id;
         } else {
             $role_id = $USER->role-id;
-        } 
-        
-        $db = DB::prepare('SELECT id FROM curriculum WHERE grade_id = ?');
-        $db->execute(array($this->id));
-        if ($db->fetchObject()){ //endroled !
-            return false;
-        } else {
-            $db = DB::prepare('DELETE FROM grade WHERE id = ?');
-            return $db->execute(array($this->id));
-        } 
+        } */
+        global $USER;
+        if (checkCapabilities('grade:delete', $USER->role_id)){
+            $db = DB::prepare('SELECT id FROM curriculum WHERE grade_id = ?');
+            $db->execute(array($this->id));
+            if ($db->fetchObject()){ //endroled !
+                return false;
+            } else {
+                $db = DB::prepare('DELETE FROM grade WHERE id = ?');
+                return $db->execute(array($this->id));
+            } 
+        }
     } 
     
     /**
